requirements.txt
discord.py
transformers
sentence-transformers
google-generativeai

# Linter and formatter tools
flake8
pylint
autoflake
isort
black
pre-commit
